+++ truffle/src/com.oracle.truffle.polyglot/src/com/oracle/truffle/polyglot/JDKSupport.java
@@ -154,7 +154,10 @@ final class JDKSupport {
             try {
                 System.load(attachLibPath);
             } catch (UnsatisfiedLinkError failedToLoad) {
-                String errorMessage = String.format("Unable to load the TruffleAttach library %s", attachLibPath);
+                String errorMessage = String.format("The Truffle attach library could not be loaded: %s. " +
+                                "This commonly happens when Truffle JARs are loaded by multiple class loaders. " +
+                                "Load polyglot languages and its dependencies from a common parent class loader, such as the application class path or module path.",
+                                attachLibPath);
                 performTruffleAttachLoadFailureAction(errorMessage, failedToLoad);
                 return Pair.create(null, errorMessage);
             } catch (IllegalCallerException illegalCaller) {